pyproject.toml
[tool.poetry]
name = "ai-content-marketing"
version = "0.1.0"
description = "A multi-agent application for automating digital content marketing. It coordinates autonomous agents to generate, optimize, and analyze content, replacing the need for a marketing team."
authors = ["Julian Lopez <jlopezsa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.1.1"
langchain = "^0.3.26"
langchain-openai = "^0.3.28"
beautifulsoup4 = "^4.13.4"
langchain-community = "^0.3.27"
langgraph = "^0.5.4"
langchain-tavily = "^0.2.11"
pillow = "^11.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
